update email verification ui
