+++ src/hooks/use-auth.ts
@@ -2,7 +2,7 @@ import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import { useCallback } from 'react'
 import constate from 'constate'
 import { useNavigate } from 'react-router-dom'
-import { fetchUser, googleSignup, login, refreshToken, resendEmail, signup } from '@/queries/auth'
+import { fetchUser, googleSignup, login, refreshToken, signup } from '@/queries/auth'
 import { deleteCookie, setCookie } from '@/utils/cookie'
 import { ACCESS_TOKEN_KEY, COOKIE_CONSENT, REFRESH_TOKEN_KEY } from '@/utils/constants'
 import { AUTH } from '@/utils/query-keys'
@@ -69,32 +69,6 @@ export function useAuth() {
     onError: (err: any) => toast(getErrorMessage(err)),
   })
 
-  const resendEmailMutation = useMutation(resendEmail, {
-    onSuccess: () => {
-      toast({
-        title: 'Email send successfully!',
-      })
-    },
-    onError: (err: any) => toast(getErrorMessage(err)),
-  })
-
-  const verifyEmailMutation = useMutation(fetchUser, {
-    onSuccess: (data) => {
-      if (!data?.email_verified) {
-        toast({
-          title: 'Email is not verified yet!',
-          variant: 'destructive',
-        })
-      } else {
-        toast({
-          title: 'Email is verified successfully!',
-        })
-        fetchUserData()
-      }
-    },
-    onError: (err: any) => toast(getErrorMessage(err)),
-  })
-
   const logout = useCallback(() => {
     deleteCookie(REFRESH_TOKEN_KEY)
     deleteCookie(ACCESS_TOKEN_KEY)
@@ -113,8 +87,6 @@ export function useAuth() {
     userData,
     fetchUserData,
     googleSignupMutation,
-    resendEmailMutation,
-    verifyEmailMutation
   }
 }
